Refactored Headcount/Afkcheck using common methods
